% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSimilarity.R
\name{compute.nbCluster.gap}
\alias{compute.nbCluster.gap}
\title{Recherche du nb de cluster par selon le critere du gap}
\usage{
compute.nbCluster.gap(val, seuil = 0, fig = FALSE)
}
\arguments{
\item{val}{#valeur propre d'une matrice de similarite}

\item{seuil}{seuil}

\item{fig}{booleen}
}
\value{
Kli
}
\description{
Recherche du nb de cluster par selon le critere du gap
}
\author{
Emilie Poisson Caillault v13/10/2015
}
