% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSimilarity.R
\name{compute.similarity.gaussien}
\alias{compute.similarity.gaussien}
\title{Calcule matrice de similarite gaussienn}
\usage{
compute.similarity.gaussien(points, sigma)
}
\arguments{
\item{points}{matrice pointsxattributs}

\item{sigma}{sigma}
}
\value{
mat
}
\description{
Calcule matrice de similarite gaussienn
}
\author{
Emilie Poisson Caillault v13/10/2015
}
