% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSimilarity.R
\name{compute.similarity.ZP}
\alias{compute.similarity.ZP}
\title{Calcule matrice de similarite gaussienne selon Zelnik-Manor et Perona}
\usage{
compute.similarity.ZP(points, vois = 7)
}
\arguments{
\item{points}{matrice pointsxattributs}

\item{vois}{nombre de voisin qui seront selectionnes}
}
\value{
mat
}
\description{
sigma local, attention risque matrice non semi-def positive
}
\author{
Emilie Poisson Caillault v13/10/2015
}
