% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalClust.R
\name{HierarchicalClust}
\alias{HierarchicalClust}
\title{Hierarchical Clustering}
\usage{
HierarchicalClust(
  W,
  K = 5,
  method = "ward.D2",
  flagDiagZero = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{W}{Gram Similarity Matrix.}

\item{K}{number of cluster to obtain.}

\item{method}{method that will be used in the hierarchical clustering.}

\item{flagDiagZero}{if True, Put zero on the similarity matrix W.}

\item{verbose}{To output the verbose in the terminal.}

\item{...}{Additional parameter for the hclust function.}
}
\value{
returns a list containing the following elements:
\itemize{
 \item{cluster: }{a vector containing the cluster}
}
}
\description{
Hierarchical Clustering
}
\examples{
### Example 1: 2 disks of the same size
n<-100 ; r1<-1
x<-(runif(n)-0.5)*2;
y<-(runif(n)-0.5)*2
keep1<-which((x*2+y*2)<(r1*2))
disk1<-data.frame(x+3*r1,y)[keep1,]
disk2 <-data.frame(x-3*r1,y)[keep1,]
sameTwoDisks <- rbind(disk1,disk2)
W <- compute.similarity.ZP(scale(sameTwoDisks))
res <- HierarchicalClust(W,K=2,method="ward.D2",flagDiagZero=TRUE,verbose=TRUE)
plot(sameTwoDisks, col = res$cluster)

### Example 2: Speed and Stopping Distances of Cars
W <- compute.similarity.ZP(scale(iris[,-5]))
res <- HierarchicalClust(W,K=2,method="ward.D2",flagDiagZero=TRUE,verbose=TRUE)
plot(iris, col = res$cluster)
}
\author{
Emilie Poisson Caillault and Erwan Vincent
}
