% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansQuantization.R
\name{kmeansQuantization}
\alias{kmeansQuantization}
\title{Data quantization}
\usage{
kmeansQuantization(dataFrame, maxData, stopCriteria = 0.99)
}
\arguments{
\item{dataFrame}{The dataFrame.}

\item{maxData}{maximum of sample number for reduction.}

\item{stopCriteria}{criterion for minimizing intra-group distance and select final smplPoint.}
}
\value{
kmeans result
}
\description{
The function use kmeans algorithm to perform data quantization.
}
\author{
Emilie Poisson Caillault and Erwan Vincent
}
