% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSimilarity.R
\name{search.neighboor}
\alias{search.neighboor}
\title{Recherche du voisin num id le plus proche}
\usage{
search.neighboor(vdist, vois)
}
\arguments{
\item{vdist}{vecteur de distance du point avec d'autres points}

\item{vois}{nombre de voisin a selectionner}
}
\value{
id
}
\description{
Recherche du voisin num id le plus proche
}
\author{
Emilie Poisson Caillault v13/10/2015
}
